调整命令中版本号部分； 增加DefaultJredic的生命周期方法； 调整dump方法返回类型； 添加一些单元测试； 添加样例模块。

+++ jredic/src/main/java/com/jredic/command/AbstractCommand.java
@@ -11,7 +11,7 @@ public abstract class AbstractCommand implements Command {
     private String[] values;
 
     //the start version value.
-    private int svv;
+    private long svv;
 
     protected AbstractCommand(String[] values, String startVersion) {
         this.values = values;
@@ -24,7 +24,7 @@ public abstract class AbstractCommand implements Command {
     }
 
     @Override
-    public int svv() {
+    public long svv() {
         return svv;
     }
